% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.choice.interactive.R
\name{cutoff.choice.interactive}
\alias{cutoff.choice.interactive}
\title{Interactive choice of cutoff points.}
\usage{
cutoff.choice.interactive(x, graph.opt = vimes.graph.opt(), ...)
}
\arguments{
\item{x}{a list of the class 'vimes.input' as returned by
\code{vimes.data}.}

\item{graph.opt}{a list of graphical options for the graphs, as
returned by \code{\link{vimes.graph.opt}}.}

\item{...}{further arguments passed to 'hist'}
}
\description{
This function displays histograms of pairwise distances and asks the user for a cut-off point to choose. It will prune each corresponding graph for any new value requested.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

