% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.choice.interactive.R
\name{cutoff_choice_interactive}
\alias{cutoff_choice_interactive}
\title{Interactive choice of cutoff points.}
\usage{
cutoff_choice_interactive(x, graph_opt = vimes_graph_opt(), ...)
}
\arguments{
\item{x}{a list of the class 'vimes_data' as returned by
\code{vimes_data}.}

\item{graph_opt}{a list of graphical options for the graphs, as
returned by \code{\link{vimes_graph_opt}}.}

\item{...}{further arguments passed to 'hist'}
}
\description{
This function displays histograms of pairwise distances and asks the user for a cut-off point to choose. It will prune each corresponding graph for any new value requested.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
