% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dconvol.R
\name{dconvol}
\alias{dconvol}
\title{Convolution of distance distributions}
\usage{
dconvol(x, type = c("t", "g", "s"), reporting.prob = 1, shape.t, scale.t,
  mu = NULL, sigma.s = NULL, tol = 0.001)
}
\arguments{
\item{x}{vector of quantiles}

\item{type}{type of distribution to be used, one of "t","g" or "s", see details}

\item{reporting.prob}{a reporting probability, between 0 and 1}

\item{shape.t}{Shape of the Gamma distribution used for the serial interval}

\item{scale.t}{of the Gamma distribution used for the serial interval}

\item{mu}{mutation rate, NULL by default as not needed for the default analysis of distance in time using only generation time parameters}

\item{sigma.s}{std of the normal distribution for spatial diffusion in each x/y direction, 
NULL by default as not needed for the default analysis of distance in time using only generation time parameters}

\item{tol}{tolerance; see vignette, the convolution accounting for underreporting uses geometric which will be cut at k.0 so that P(X>=k.0) < tol.}
}
\value{
a list of dist objects with matching entries, with attributes: 'labels' (labels of the cases) and 'N' (number of cases)
}
\description{
This function computes the probability density function of convoolution of various distance distributions, representing distance between cases in time, space or genetic.
}
\details{
\code{type} can be one of "t","g" or "s". 
For \code{type} "t", temporal convolution is performed assuming Gamma distributed serial interval.
For \code{type} "g", genetic distance convolution is performed assuming NegBin distribution 
(convolution of Gamma distributed serial interval and Poisson distributed number of mutations).
For \code{type} "s", spatial Euclidian distance is performed assuming Rayleigh distribution.
}
\examples{

}
\author{
Anne Cori \email{a.cori@imperial.ac.uk}
}

