% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpaircase.R
\name{fpaircase}
\alias{fpaircase}
\alias{print.fpaircase}
\alias{plot.fpaircase}
\title{Expected distributions of distances (functions)}
\usage{
fpaircase(type = c("temporal", "genetic", "spatial", "empiric"), gamma_shape,
  gamma_rate = 1, gamma_scale = 1/gamma_rate, poisson_rate, sd_spatial, p,
  alpha = 0.001)

\method{print}{fpaircase}(x, ...)

\method{plot}{fpaircase}(x, y = NULL, xlim = c(0, 10), ...)
}
\arguments{
\item{type}{type of distance to be considered (one of "temporal","genetic",
"spatial" or "empiric").}

\item{sd_spatial}{standard deviation of the Normal spatial kernel.}

\item{alpha}{The probability threshold to be used to determine the maximum
value of generations between two successive cases to consider;
this value ('max_kappa') will be the smallest k so that
p(k > max_kappa) < alpha. Defaults to 0.001.}

\item{x}{vector of quantiles.}

\item{...}{further arguments passed to methods (print, plot, etc.)}
}
\description{
The function \code{fpaircase} returns functions which compute the expected
distributions of distances between pairs of cases given a case reporting
probability 'pi'. See \code{\link{dpaircase}} for details on different types
of distances between cases for which distributions can be computed.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}.
}
