% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quantiles.R
\name{get_quantiles_pdf}
\alias{get_quantiles_pdf}
\alias{get_quantiles}
\alias{get_quantiles_pmf}
\title{Get quantiles for a given probability distribution}
\usage{
get_quantiles_pdf(f, p, precision = 0.01, n_steps = 1000, ...)

get_quantiles_pmf(f, p, n_steps = 1000, ...)
}
\arguments{
\item{f}{A function returning the probability density function (for
\code{get_quantiles_pdf}) or the probability mass function (for
\code{get_quantiles_pmf}.}

\item{p}{A vector of probabilities.}

\item{precision}{The size of the step used to discretise the continuous
distribution. Defaults to 0.01.}

\item{n_steps}{The number steps used to discretise the continuous
distribution. Defaults to 1000. Note that if \code{n_steps} intervals are
not enough to attain the largest value in \code{p}, batches of
\code{n_steps} are iteratively added.}
}
\description{
These functions approximate quantiles for a probability density function or
mass function given by a function \code{f}.
}
\examples{

## reference: exponential of rate 1
qexp(c(.4, .5, .95))

## approximation
get_quantiles_pdf(f = dexp, c(.4,.5,.95))

## better precision
get_quantiles_pdf(f = dexp, c(.4,.5,.95), precision = 0.001)

}
\author{
Anne Cori (a.cori@imperial.ac.uk) and Thibaut Jombart.
}
