% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimes.pal.R
\name{vimes.pal1}
\alias{palettes}
\alias{vimes.pal1}
\alias{vimes.pal2}
\title{Color palettes used in vimes}
\usage{
vimes.pal1(n)

vimes.pal2(n)
}
\arguments{
\item{n}{a number of colors}
}
\description{
These functions are color palettes used in vimes.
}
\examples{

plot(1:4, cex=8, pch=20, col=vimes.pal1(4), main="vimes.pal1")
plot(1:20, col=vimes.pal1(20), pch=20, cex=6, main="vimes.pal1")
plot(1:10, col=vimes.pal2(10), pch=20, cex=6, main="vimes.pal2")
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

