% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vimes.input.R
\name{plot.vimes.input}
\alias{plot.vimes.input}
\title{Plot input data for vimes}
\usage{
\method{plot}{vimes.input}(x, y = NULL, col = vimes.pal1(length(x)),
  n = length(x), which = 1:length(x), ...)
}
\arguments{
\item{x}{a \code{vimes.input} object, as returned by \code{vimes.dist}}

\item{y}{left for compatibility with the generic 'plot'}

\item{col}{colors to be used for the histograms}

\item{n}{the number of histograms displayed per window}

\item{which}{an integer vector indicating which distances to plot}

\item{...}{further arguments passed on to \code{hist}}
}
\description{
This function makes a series of histograms for data input to vimes.
}
\examples{
x1 <- rnorm(20)
x2 <- runif(20)
names(x1) <- sample(letters, 20)
names(x2) <- sample(letters, 20)
D1 <- dist(x1)
D2 <- dist(x2)
out <- vimes.dist(AnkhDist=D1, MorpokDist=D2)
out
plot(out)

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

