% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimes.data.R
\name{vimes.data}
\alias{vimes.data}
\title{Process input data for vimes}
\usage{
vimes.data(dates = NULL, xy = NULL, dna = NULL, lonlat = FALSE, ...)
}
\arguments{
\item{dates}{a vector of dates for each cases; typically dates of onset are used; the vector
should be named if elements do not match perfectly the other data types}

\item{xy}{a 2-columns matrix or data.frame containing locations of the cases; if \code{lonlat} is
\code{TRUE}, coordinates are taken as longitudes and latitudes and the great circle distance will
be used; rows should be named if elements do not match perfectly the other data types}

\item{dna}{a matrix with class \code{DNAbin} (from the package 'ape'); rows should be named if
elements do not match perfectly the other data types}

\item{...}{other data to be used foir computing pairwise distances using \code{dist}; entries
should be named if elements do not match perfectly the other data types; \code{dist} objects are
left as is.}
}
\value{
a list of named dist objects with matching entries, with attributes: 'labels' (labels of
the cases) and 'N' (number of cases); distances based on dates, xy coordinates, and genetic data
will be named 'temporal', 'spatial' and 'genetic', respectively
}
\description{
This function processes input data for vimes, computing for each data a matrix of pairwise distances.
Data can have the following pre-defined types:
\itemize{
\item{dates}{dates associated to cases, typically dates of onset of symptoms}
\item{spatial}{matrix or data.frame with two columns for geographic coordinates}
\item{DNA sequences}{aligned DNA sequences of the pathogen}
}
}
\details{
Data without a pre-defined type will be fed directly to the function \code{dist}.
Objects which are already pairwise distances with the class \code{dist} are left as is.
}
\examples{
x <- vimes.data(dates=sim1$dates, xy=sim1$xy, dna=sim1$dna)
x
plot(x)

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
\code{\link{vimes.dist}} for further processing of the resulting pairwise distance
matrices; internally called by \code{vimes.data}
}

