% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimes.pal.R
\name{vimes.pal1}
\alias{vimes.pal1}
\title{Color palettes used in vimes}
\usage{
vimes.pal1(n)
}
\arguments{
\item{n}{a number of colors}
}
\description{
These functions are color palettes used in vimes.
}
\examples{

plot(1:8, cex=8, pch=20, col=vimes.pal1(8), main="vimes.pal1")

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

