% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimes.prune.ml.R
\name{vimes.prune.ml}
\alias{vimes.prune.ml}
\title{Builds a graph from pairwise distances, optimizing cutoff using ML}
\usage{
vimes.prune.ml(x, f, graph.opt = vimes.graph.opt())
}
\arguments{
\item{x}{pairwise distances stored as a \code{dist} object.}

\item{f}{a function giving the log-likelihood of a given distance between two epidemiologically conected cases.}

\item{graph.opt}{a list of graphical options for the graphs, as returned by \code{\link{vimes.graph.opt}}.}
}
\value{
a list containing:
\describe{
\item{graph}{a graph connecting individuals, weighted with the input distances, with \code{igraph} class (from the \code{igraph} package).}
\item{clusters}{a list providing cluster information: group membership, cluster sizes, and the number of clusters (K)}
}
}
\description{
This function builds a graph using a \code{dist} object and a cut-off distance beyond which individuals are no longer connected.
The cutoff distance is chosen using Maximum-Likelihood estimation. The likelihood function has to be provided.
}
\details{
The cutoff is inclusive: only cases strictly further away than the cutoff distance will be disconnected on the graph. Note that this differs from the original \code{gengraph} implementation in the package \code{adegenet}.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
the function \code{gengraph} in the package \code{adegenet}, which was an initial implementation of the same idea in a genetics  context.
}

