% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimes_data.R
\name{vimes_data}
\alias{vimes_data}
\title{Process input data for vimes}
\usage{
vimes_data(...)
}
\arguments{
\item{...}{a series (alternatively, a list) of matrices or dist objects serving as input.}
}
\value{
a list of dist objects with matching entries, with attributes: 'labels' (labels of the cases) and 'N' (number of cases)
}
\description{
This function takes a series of 'dist' objects (or objects which can be converted to 'dist', and ensures that entries match in all objects, adding NAs where necessary. The total number of cases and labels are returned as attributes.
}
\examples{
x1 <- c(0,1,3)
x2 <- c(2,5)
names(x1) <- letters[1:3]
names(x2) <- c('a', 'r')
D1 <- dist(x1)
D2 <- dist(x2)
out <- vimes_data(D1, D2)
out

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
