% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dpaircase}
\alias{dempiric}
\alias{dgenetic}
\alias{dpaircase}
\alias{dspatial}
\alias{dtemporal}
\title{Expected distributions of distances}
\usage{
dpaircase(pi, alpha = 0.001)

dtemporal(x, shape, rate = 1, scale = 1/rate, pi, alpha = 0.001)

dspatial(x, sd, pi, alpha = 0.001)

dgenetic(x, gamma_shape, gamma_rate = 1, gamma_scale = 1/gamma_rate,
  poisson_rate, pi, alpha = 0.001)

dempiric(p, pi, alpha = 0.001)
}
\arguments{
\item{pi}{The reporting probability, i.e. the proportion of cases of the
outbreak that have been reported.}

\item{alpha}{The probability threshold to be used to determine the maximum
value of generations between two successive cases to consider; 
this value ('max_kappa') will be the smallest k so that
p(k > max_kappa) < alpha. Defaults to 0.001.}

\item{x}{vector of quantiles.}

\item{shape, scale}{shape and scale of the gamma distribution used for the serial interval}

\item{rate}{an alternative way to specify the scale of the gamma distribution used for the serial interval}

\item{sd}{standard deviation of the Normal spatial kernel.}

\item{gamma_shape, gamma_scale}{shape and scale of the gamma distribution used for the serial interval}

\item{gamma_rate}{an alternative way to specify the scale of the gamma distribution used for the serial interval}

\item{poisson_rate}{rate (i.e. mean) of the poisson distribution used for the per generation genetic mutation rate}

\item{p}{A \code{numeric} vector providing the probability mass function, or
empirical frequencies, of pairwise distances.}
}
\description{
These functions compute the expected distributions of distances between pairs
of cases given a case reporting probability 'pi'. Analytical results are used
for some special cases, including:
}
\details{
\itemize{

\item temporal distances: the serial interval is assumed to be
gamma-distributed; the procedure returns a discretized, weighted convolution
of gamma distributions.

\item genetic distances: assumed ... TBC

}
}
\author{
Anne Cori (a.cori@imperial.ac.uk) and Thibaut Jombart
(thibautjombart@gmail.com).
}

