% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dpaircase}
\alias{dpaircase}
\alias{dtemporal}
\alias{dspatial}
\alias{dgenetic}
\alias{dempiric}
\title{Expected distributions of distances}
\usage{
dpaircase(pi, alpha = 0.001)

dtemporal(pi, alpha = 0.001)

dspatial(pi, alpha = 0.001)

dgenetic(pi, alpha = 0.001)

dempiric(x, pi, alpha = 0.001)
}
\arguments{
\item{pi}{The reporting probability, i.e. the proportion of cases of the
outbreak that have been reported.}

\item{alpha}{The probability threshold to be used to determine the maximum
value of kappa; this value ('max_kappa') will be the smallest kappa so that
p(kappa > max_kappa) < alpha. Defaults to 0.001.}

\item{x}{A \code{numeric} vector providing the probability mass function, or
empirical frequencies, of pairwise distances.}
}
\description{
These functions compute the expected distributions of distances between pairs
of cases given a case reporting probability 'pi'. Analytical results are used
for some special cases, including:
}
\details{
\itemize{

\item temporal distances: the serial interval is assumed to be
gamma-distributed; the procedure returns a discretized, weighted convolution
of gamma distributions.

\item genetic distances: assumed ... TBC

}
}
\author{
Anne Cori (a.cori@imperial.ac.uk) and Thibaut Jombart
(thibautjombart@gmail.com).
}
