% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dpaircase}
\alias{dempiric}
\alias{dgenetic}
\alias{dpaircase}
\alias{dspatial}
\alias{dtemporal}
\title{Expected distributions of distances}
\usage{
dpaircase(pi, alpha = 0.001)

dtemporal(x, shape, rate = 1, scale = 1/rate, pi, alpha = 0.001)

dspatial(x, sd, pi, alpha = 0.001)

dgenetic(x, gamma_shape, gamma_rate = 1, gamma_scale = 1/gamma_rate,
  poisson_rate, pi, alpha = 0.001)

dempiric(p, pi, alpha = 0.001)
}
\arguments{
\item{pi}{The reporting probability, i.e. the proportion of cases of the
outbreak that have been reported.}

\item{alpha}{The probability threshold to be used to determine the maximum
value of generations between two successive cases to consider; 
this value ('max_kappa') will be the smallest k so that
p(k > max_kappa) < alpha. Defaults to 0.001.}

\item{x}{vector of quantiles.}

\item{shape, scale}{shape and scale of the gamma distribution used for the serial interval}

\item{rate}{an alternative way to specify the scale of the gamma distribution used for the serial interval}

\item{sd}{standard deviation of the Normal spatial kernel.}

\item{gamma_shape, gamma_scale}{shape and scale of the gamma distribution used for the serial interval}

\item{gamma_rate}{an alternative way to specify the scale of the gamma distribution used for the serial interval}

\item{poisson_rate}{rate (i.e. mean) of the poisson distribution used for the per generation genetic mutation rate}

\item{p}{A \code{numeric} vector providing the probability mass function, or
empirical frequencies, of pairwise distances.}
}
\description{
These functions compute the expected distributions of distances between pairs
of cases given a case reporting probability 'pi'. Analytical results are used
for some special cases, including:
}
\details{
\itemize{

\item temporal distances: the serial interval is assumed to be
gamma-distributed; the procedure returns a discretized, weighted convolution
of gamma distributions.

\item genetic distances: assumed ... TBC

}
}
\examples{
 #### compare dempiric and dtemporal ####
 
 ## compute empirical distribution correponding to exponential(mean 50)
 mean_exp <- 50
 x <- 1:300
 reporting_rate <- 0.5
 p <- dgamma(x, shape=mean_exp, rate=1) 
       # computes pdf of a gamma distr with shape mean_exp and scale=rate=1 
       # (i.e. an exponential distr with mean mean_exp)

 ## use this as an empirical distribution to feed into dempiric
 empiric_exp_distr_with_underreporting <- dempiric(p, pi=reporting_rate)
 temporal_distr_with_underreporting <- dtemporal(x, 
     shape=mean_exp, rate=1, pi=reporting_rate)
 
 ## compare the two
 correlation <- cor(empiric_exp_distr_with_underreporting[1:length(x)],
     temporal_distr_with_underreporting)
 
 ## graphical comparison
 plot(x, empiric_exp_distr_with_underreporting[1:length(x)], xlab="Time", ylab="pdf", 
       main="Pdf of time between a case and its closest ancestry in dataset
             when SI is exponentially distributed with mean 50,
             and reporting probability is 0.5", cex.main=1, pch=3)
 lines(x, temporal_distr_with_underreporting, col="red")
 legend("topright",c("using dempiric","using dtemporal"), 
     pch=c(3, -1), lwd=c(-1, 1), col=c("black","red"), bty="n")
}
\author{
Anne Cori (a.cori@imperial.ac.uk) and Thibaut Jombart
(thibautjombart@gmail.com).
}

