% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimes.pal.R
\name{vimes_pal1}
\alias{vimes_pal1}
\alias{palettes}
\alias{vimes_pal2}
\alias{vimes_pal2}
\title{Color palettes used in vimes}
\usage{
vimes_pal1(n)

vimes_pal2(n)
}
\arguments{
\item{n}{a number of colors}
}
\description{
These functions are color palettes used in vimes.
}
\examples{

plot(1:4, cex=8, pch=20, col=vimes_pal1(4), main="vimes_pal1")
plot(1:20, col=vimes_pal1(20), pch=20, cex=6, main="vimes_pal1")
plot(1:10, col=vimes_pal2(10), pch=20, cex=6, main="vimes_pal2")
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
