% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimes.data.R
\name{vimes.data}
\alias{vimes.data}
\title{Check input data}
\usage{
vimes.data(...)
}
\arguments{
\item{...}{a set of matrices or dist objects serving as input}
}
\description{
Check input data
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

