% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimes.data.R
\name{vimes.data}
\alias{vimes.data}
\title{Process input data for vimes}
\usage{
vimes.data(...)
}
\arguments{
\item{...}{a series (alternatively, a list) of matrices or dist objects serving as input.}
}
\value{
a list of dist objects with matching entries, with attributes: 'labels' (labels of the cases) and 'N' (number of cases)
}
\description{
This function processes input data for vimes, computing for each data a matrix of pairwise distances.
Data can have the following pre-defined types:
\itemize{
\item{dates}{dates associated to cases, typically dates of onset of symptoms}
\item{spatial}{matrix or data.frame with two columns for geographic coordinates}
\item{DNA sequences}{aligned DNA sequences of the pathogen}
}
}
\details{
Data without a pre-defined type will be fed directly to the function \code{dist}.
Objects which are already pairwise distances with the class \code{dist} are left as is.
}
\examples{
x1 <- c(0,1,3)
x2 <- c(2,5)
names(x1) <- letters[1:3]
names(x2) <- c('a', 'r')
D1 <- dist(x1)
D2 <- dist(x2)
out <- vimes.data(D1, D2)
out

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
\code{\link{vimes.dist}} for further processing of the resulting pairwise distance
matrices; internally called by \code{vimes.data}
}

